requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==5.2
django-allauth==65.7.0
django-contact-form==5.2.0
django-environ==0.12.0
django-filter==25.1
django-model-utils==5.0.0
django-private-chat2==1.0.2
idna==3.10
msgpack==1.1.0
pillow==11.2.1
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.2
urllib3==2.4.0
